payment-service/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9
PyJWT>=2.9
qrcode[pil]>=7.4